backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic-core==2.23.4
python-dotenv==1.0.1
